pyproject.toml
[project]
name = "tts-to-obsidian-markdown"
version = "0.1.0"
description = "Convert voice recordings to structured Obsidian diary entries"
authors = [
    {name = "jxsh dxncxn", email = "pythonicqualms@gmail.com"}
]
dependencies = [
    "PyAudio>=0.2.13",
    "sounddevice>=0.4.6",
    "numpy>=1.24.0",
    "requests>=2.31.0",
    "PyYAML>=6.0.1",
    "python-dateutil>=2.8.2",
    "nltk>=3.8.1",
    "spacy>=3.7.2",
    "pydub>=0.25.1",
    "watchdog>=3.0.0",
    "rich>=13.7.0",
    "typer>=0.9.0",
    "python-dotenv>=1.0.0",
    "pip>=25.1.1",
    "openai-whisper>=20240930",
]

[tool.ruff]
line-length = 88
target-version = "py39"

[tool.ruff.lint]
select = ["E", "F", "B", "I"]
ignore = []

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"
